#filmorate !11 refactoring: разбил тесты на классы как и должно было быть изначально
